# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.14"
  title: DTL Service
# during dev, should point to your local machine
host: localhost:80
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
  - application/octet-stream #smell, this is necessary because supertest appears to send this content type for empty posts
# format of the responses to the client (Accepts)
produces:
  - application/json
  - text/plain
  - text/html
paths:
...
  /findUsersWithAllianceMeta/users:
    x-swagger-router-controller: PlayerController
    post:
      description: Find Users with Alliance Metadata by list of User IDs
      operationId: findUsersWithAllianceMeta
      parameters:
        - name: body
          in: body
          description: JSON document that has teams and their related summonerIds and metadata
          required: true
          schema:
            type: object
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/EntityResponse"
        "default":
          description: Failure
          schema:
            $ref: "#/definitions/ErrorResponse"
  /findAllUserAlliances/alliances:
    x-swagger-router-controller: PlayerController
    post:
      description: Find User-Alliances by list of Alliance IDs
      operationId: findAllUserAlliances
      parameters:
        - name: body
          in: body
          description: JSON document that has teams and their related summonerIds and metadata
          required: true
          schema:
            type: object
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/UserAlliancesResponse"
        "default":
          description: Failure
          schema:
            $ref: "#/definitions/ErrorResponse"
## atomic entity operations
  # /test:
  #   x-swagger-router-controller: TestController
  #   get:
  #     description: test
  #     operationId: migrateUp
  #     responses:
  #       "200":
  #         description: success
...
  /playerbyuserandseries/user/{slgUserId}/series/{seriesId}:
    x-swagger-router-controller: PlayerController
    get:
      description: Get players by User ID and Series ID
      operationId: getPlayerByUserAndSeries
      parameters:
      - name: slgUserId
        type: number
        in: path
        required: true
        description: The unqiue numerical identifier for User
      - name: seriesId
        type: number
        in: path
        required: true
        description: The unqiue numerical identifier for Series
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/PlayerListResponse"
        "default":
          description: Failure
          schema:
            $ref: "#/definitions/ErrorResponse"
  # /allianceStandings/experience/{experienceId}:
  #   x-swagger-router-controller: AllianceController
  #   get:
  #     description: Get Alliance Standings by Experience ID
  #     operationId: getAllianceStangings
  #     parameters:
  #     - name: experienceId
  #       type: number
  #       in: path
  #       required: true
  #       description: The unqiue numerical identifier for Experience
  #     responses:
  #       "200":
  #         description: Success
  #         schema:
  #           $ref: "#/definitions/AllianceStandingsResponse"
  #       "default":
  #         description: Failure
  #         schema:
  #           $ref: "#/definitions/ErrorResponse"
...
  /findIndPlayerStanding/experience/{experienceId}/series/{seriesId}/event/{eventId}:
    x-swagger-router-controller: PlayerController
    get:
      description: Get Player Standings by Experience, Series, and Event ID
      operationId: getIndPlayerStanding
      parameters:
      - name: experienceId
        type: number
        in: path
        required: true
        description: The unqiue numerical identifier for Experience
      - name: seriesId
        type: number
        in: path
        required: true
        description: The unqiue numerical identifier for Series
      - name: eventId
        type: number
        in: path
        required: true
        description: The unqiue numerical identifier for Event
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/PlayerStandingsResponse"
        "default":
          description: Failure
          schema:
            $ref: "#/definitions/ErrorResponse"
  /findIndPlayerStandingCustomQry/experience/{experienceId}/series/{seriesId}/event/{eventId}:
    x-swagger-router-controller: PlayerController
    get:
      description: Get Player Standings by Experience, Series, and Event ID
      operationId: getIndPlayerStandingCustomQry
      parameters:
      - name: experienceId
        type: number
        in: path
        required: true
        description: The unqiue numerical identifier for Experience
      - name: seriesId
        type: number
        in: path
        required: true
        description: The unqiue numerical identifier for Series
      - name: eventId
        type: number
        in: path
        required: true
        description: The unqiue numerical identifier for Event
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/PlayerStandingsResponse"
        "default":
          description: Failure
          schema:
            $ref: "#/definitions/ErrorResponse"
  /findIndPlayerStanding/experience/{experienceId}/series/{seriesId}:
    x-swagger-router-controller: PlayerController
    get:
      description: Get Player Standings by Experience and Series ID
      operationId: getIndPlayerStanding
      parameters:
      - name: experienceId
        type: number
        in: path
        required: true
        description: The unqiue numerical identifier for Experience
      - name: seriesId
        type: number
        in: path
        required: true
        description: The unqiue numerical identifier for Series
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/PlayerStandingsResponse"
        "default":
          description: Failure
          schema:
            $ref: "#/definitions/ErrorResponse"
  /findIndPlayerStanding/experience/{experienceId}:
    x-swagger-router-controller: PlayerController
    get:
      description: Get Player Standings by Experience ID
      operationId: getIndPlayerStanding
      parameters:
      - name: experienceId
        type: number
        in: path
        required: true
        description: The unqiue numerical identifier for Experience
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/PlayerStandingsResponse"
        "default":
          description: Failure
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /findMonthlySeriesOptions/experience/{experienceId}/series/{seriesId}/event/{eventId}:
    x-swagger-router-controller: PlayerController
    get:
      description: Find Series Options by Experience, Series, and Event ID
      operationId: findSeriesOptions
      parameters:
      - name: experienceId
        type: number
        in: path
        required: true
        description: The unqiue numerical identifier for Experience
      - name: seriesId
        type: number
        in: path
        required: true
        description: The unqiue numerical identifier for Series
      - name: eventId
        type: number
        in: path
        required: true
        description: The unqiue numerical identifier for Event
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/SeriesResponse"
        "default":
          description: Failure
          schema:
            $ref: "#/definitions/ErrorResponse"
  # TODO: rename since montly series concept is from v1 services
  /findMonthlySeriesOptions/experience/{experienceId}/series/{seriesId}:
    x-swagger-router-controller: PlayerController
    get:
      description: Find Series Options by Experience and Series ID
      operationId: findSeriesOptions
      parameters:
      - name: experienceId
        type: number
        in: path
        required: true
        description: The unqiue numerical identifier for Experience
      - name: seriesId
        type: number
        in: path
        required: true
        description: The unqiue numerical identifier for Series
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/SeriesResponse"
        "default":
          description: Failure
          schema:
            $ref: "#/definitions/ErrorResponse"
# TODO: rename since montly series concept is from v1 services
  /findMonthlySeriesOptions/experience/{experienceId}:
    x-swagger-router-controller: PlayerController
    get:
      description: Find Series Options by Experience ID
      operationId:  findSeriesOptions
      parameters:
      - name: experienceId
        type: number
        in: path
        required: true
        description: The unqiue numerical identifier for Experience
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/SeriesResponse"
        "default":
          description: Failure
          schema:
            $ref: "#/definitions/ErrorResponse"
  /individualstandings/experience/{experienceId}/series/{seriesId}/event/{eventId}:
    x-swagger-router-controller: PlayerController
    get:
      description: Find Individual Standings by Experience, Series, and Event ID
      operationId: getIndPlayerStanding
      parameters:
      - name: experienceId
        type: number
        in: path
        required: true
        description: The unqiue numerical identifier for Experience
      - name: seriesId
        type: number
        in: path
        required: true
        description: The unqiue numerical identifier for Series
      - name: eventId
        type: number
        in: path
        required: true
        description: The unqiue numerical identifier for Event
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/IndividualStandingsResponse"
        "default":
          description: Failure
          schema:
            $ref: "#/definitions/ErrorResponse"
  /individualstandings/experience/{experienceId}/series/{seriesId}:
    x-swagger-router-controller: PlayerController
    get:
      description: Find Individual Standings by Experience and Series ID
      operationId: getIndPlayerStanding
      parameters:
      - name: experienceId
        type: number
        in: path
        required: true
        description: The unqiue numerical identifier for Experience
      - name: seriesId
        type: number
        in: path
        required: true
        description: The unqiue numerical identifier for Series
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/IndividualStandingsResponse"
        "default":
          description: Failure
          schema:
            $ref: "#/definitions/ErrorResponse"
  /individualstandings/experience/{experienceId}:
    x-swagger-router-controller: PlayerController
    get:
      description: Find Individual Standings by Experience ID
      operationId: getIndPlayerStanding
      parameters:
      - name: experienceId
        type: number
        in: path
        required: true
        description: The unqiue numerical identifier for Experience
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/IndividualStandingsResponse"
        "default":
          description: Failure
          schema:
            $ref: "#/definitions/ErrorResponse"
  /findAllPlayers/experience/{experienceId}:
    x-swagger-router-controller: PlayerController
    get:
      description: Find Players by Experience ID
      operationId: findAllPlayers
      parameters:
      - name: experienceId
        type: number
        in: path
        required: true
        description: The unqiue numerical identifier for Experience
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/PlayerListResponse"
        "default":
          description: Failure
          schema:
            $ref: "#/definitions/ErrorResponse"
  /findAllPlayersWithMeta/experience/{experienceId}:
    x-swagger-router-controller: PlayerController
    get:
      description: Find Players with Metadata by Experience ID
      operationId: findAllPlayersWithMeta
      parameters:
      - name: experienceId
        type: number
        in: path
        required: true
        description: The unqiue numerical identifier for Experience
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/PlayersWithMetaResponse"
        "default":
          description: Failure
          schema:
            $ref: "#/definitions/ErrorResponse"          
# word press schema operations
...
  /playerbyuserseries/{userId}/{seriesId}:
    x-swagger-router-controller: PlayerController
    get:
      description: Assign user to alliance
      operationId: findPlayerByUserSeries
      parameters:
      - name: userId
        type: number
        in: path
        required: true
        description: The unqiue numerical identifier for User
      - name: seriesId
        type: number
        in: path
        required: true
        description: The unqiue numerical identifier for Series
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Player"
        "default":
          description: Failure
          schema:
            $ref: "#/definitions/ErrorResponse"         
  # /eventsbyseries/series/{seriesId}:
  #   x-swagger-router-controller: ORMController
  #   get:
  #     description: Get all entities
  #     operationId: findAllEntities
  #     parameters:
  #       - name: seriesId
  #         type: number
  #         in: path
  #         required: true
  #         description: entity name
  #     responses:
  #       "200":
  #         description: Success
  #         schema:
  #           $ref: "#/definitions/EntityResponse"
  #       "default":
  #         description: Failure
  #         schema:
  #           $ref: "#/definitions/ErrorResponse"           
# complex objects have schema definitions
definitions:
...
  EntityResponse:
    type: object
  # Response definitions
  ErrorResponse:
    required:
      - type
      - code
      - message
    properties:
      type:
        type: string
      code:
        type: number
      message:
        type: string
...
  PlayerListResponse:
    type: array
    items: 
      $ref: "#/definitions/Player"
...
  AllianceStandingsResponse:
    type: array
    items: 
      $ref: "#/definitions/AllianceStanding"
  PlayerStandingsResponse:
    type: array
    items: 
      $ref: "#/definitions/PlayerStanding"
  # TODO: rename since montly series concept is from v1 services
  SeriesResponse:
    required:
      - optionsData
      - selected
    properties:
      # TODO: rename since weekly series concept is from v1 services
      # The array of weekly series options
      optionsData:
        type: array
        items:
          $ref: "#/definitions/EventOption"
      selected:
        $ref: "#/definitions/SeriesOption"
  IndividualStandingsResponse:
    type: array
    items: 
      $ref: "#/definitions/IndividualStanding"
  PlayersWithMetaResponse:
    type: array
    items: 
      $ref: "#/definitions/PlayerWithMeta"
...
  UserAlliancesResponse:
    type: array
    items: 
      $ref: "#/definitions/UserAlliance"
  # Base definitions
...
  AllianceStanding:
    type: object
...
  IndividualStanding:
    type: object
  Player:
    type: object
  PlayerStanding:
    type: object
  UserAlliance:
    type: object
  # TODO: probably need to be more specific about the included metadata
  PlayerWithMeta:
    type: object
...
  SeriesOption:
    type: object
...
  EventOption:
    required:
      - optionsData
      - selected
    properties:
      # The array of series options
      optionsData:
        type: array
        items:
          $ref: "#/definitions/SeriesOption"
      selected:
        type: boolean
# complex objects have schema definitions